Fix IDE hang when projects with different "Process files asynchronously" are open
Fixes #160

This problem only happens when at least 2 projects with different "Process files asynchronously" are open and a save is executed.

If a save (Ctrl + S) is performed it saves both projects in parallel, causing a lock failure as one project locks the UI while the other one (which has the lock / runs async) tries to acquire the UI thread, thus causing a deadlock.